+++ event-sourcing/event-source-api/src/main/java/uk/gov/justice/services/eventsourcing/discovery/EventSubscriptionDiscoverer.java
@@ -5,5 +5,5 @@ import java.util.UUID;
 
 public interface EventSubscriptionDiscoverer {
 
-    DiscoveryResult discoverNewEvents(Optional<UUID> latestKnownEventId, int batchSize);
+    DiscoveryResult discoverNewEvents(Optional<UUID> latestKnownEventId, int batchSize, String source);
 }