+++ event-sourcing/event-source-api/src/main/java/uk/gov/justice/services/eventsourcing/eventreader/RestNextEventReaderException.java
@@ -1,0 +1,12 @@
+package uk.gov.justice.services.eventsourcing.eventreader;
+
+public class RestNextEventReaderException extends RuntimeException {
+
+    public RestNextEventReaderException(final String message) {
+        super(message);
+    }
+
+    public RestNextEventReaderException(final String message, final Throwable cause) {
+        super(message, cause);
+    }
+}